RandStrParser, _write_file(), and main() moved to own file, randstr_terminal.py
randstr's core class was seperated from randstr's terminal integration classes and functions. This seperation improves organization, and it makes seperating all the imports much easier

Docstring explaining how RandomString accesses SystemRandom was also
reduced as the README already covers this in detail
